os/bluestore: change the type of aio_t:res to long
io_event::res was assigned to io_t::rval, but io_event::res
is long, if rval is int, this will cause overflow.

Fixes: http://tracker.ceph.com/issues/23527

(cherry picked from commit a9b9477a44d4556f6827e6dc4b94a5f1f61ce8e1)

+++ src/os/bluestore/KernelDevice.cc
@@ -370,7 +370,7 @@ void KernelDevice::_aio_thread()
 	// later flush() occurs.
 	io_since_flush.store(true);
 
-	int r = aio[i]->get_return_value();
+	long r = aio[i]->get_return_value();
         if (r < 0) {
           derr << __func__ << " got " << cpp_strerror(r) << dendl;
           if (ioc->allow_eio && r == -EIO) {
